+++ components/admin/Blog/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+interface Props {}
+
+const AdminBlog = ({}: Props) => {
+  return <h1>Admin Blog</h1>;
+};
+
+export default AdminBlog;