+++ data/consts.ts
@@ -2,6 +2,19 @@ export const LOGIN_EXPIRY_SECONDS = 86400;
 
 export const DEFAUlT_IMG = 'default.png';
 
+export const ROUTES = {
+  ADMIN: '/admin?section=home',
+  HOME: '/',
+};
+
+export const ADMIN_SECTIONS = {
+  HOME: '?section=home',
+  BLOG_HOME: '?section=blog',
+  BLOG_ADD: '?section=blog-add',
+  BLOG_EDIT: '?section=blog-edit',
+  USER: '?section=user',
+};
+
 export const PAGE_LAYERS = {
   TOP_VIEW: 5000,
   OVERLAY: 4000,