+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Changes will be documented here generally around release items using symantic ve
 ## Unreleased
 
 - [login-page] working on designing the login page
+- [admin] building out the admin section of the site
+  - create subdomains
 
 ## Version 1.0.0